requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appnope==0.1.0            # via ipython
backcall==0.1.0           # via ipython
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
coverage==4.5.4
decorator==4.4.0          # via ipython, traitlets
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-environ==0.4.5
django-guardian==2.0.0
django-ical==1.5
django-recaptcha2==1.4.1
django-rest-swagger==2.2.0
django==2.2.3
djangorestframework==3.10.2
future==0.17.1
gitdb2==2.0.5             # via gitpython
gitpython==2.1.13
gunicorn==19.9.0
icalendar==4.0.3
idna==2.8                 # via requests
ipython-genutils==0.2.0   # via traitlets
ipython==7.7.0
itypes==1.1.0             # via coreapi
jedi==0.14.1              # via ipython
jinja2==2.10.1            # via coreschema
markupsafe==1.1.1         # via jinja2
openapi-codec==1.3.2      # via django-rest-swagger
parso==0.5.1              # via jedi
passlib==1.7.1
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==3.8.0
prompt-toolkit==2.0.9     # via ipython
psycopg2-binary==2.8.3
ptyprocess==0.6.0         # via pexpect
pygments==2.4.2           # via ipython
python-dateutil==2.8.0    # via icalendar
pytz==2019.1
requests==2.22.0          # via coreapi, django-recaptcha2
simplejson==3.16.0        # via django-rest-swagger
six==1.12.0               # via pip-tools, prompt-toolkit, python-dateutil, traitlets
smmap2==2.0.5             # via gitdb2
sqlparse==0.3.0           # via django, django-debug-toolbar
traitlets==4.3.2          # via ipython
uritemplate==3.0.0        # via coreapi
urllib3==1.25.3           # via requests
wcwidth==0.1.7            # via prompt-toolkit
whitenoise==4.1.3

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via ipython
